cts urns adjusted
Removed leading zeros in document identifiers to match cts urn patterns. Each work should be treated as a distinct document because they are not based on a manuscript exemplar and because of the radical differences among reconstructions of a lost text, and work titles should include the editor's/reconstructor's name as their original creation. There are no existing TLG document identifiers for Marcion's Gospel or Marcion's Apostle in use in other corpora (e.g., TLG), so there should not be any comflict. Essentially, our team is minting new document identifiers.